+++ snd-x1.h
@@ -439,7 +439,7 @@ void g_init_gxsnd(void);
 
 /* -------- snd-xfile.c -------- */
 
-char *read_file_data_choices(file_data *fdat, int *srate, int *chans, int *type, int *format, int *location);
+char *read_file_data_choices(file_data *fdat, int *srate, int *chans, int *type, int *format, off_t *location);
 file_data *make_file_data_panel(snd_state *ss, Widget parent, char *name, Arg *args, int n, int with_chan, 
 				int header_type, int data_format, int with_loc, int with_comment);
 void alert_new_file(void);
