+++ snd-xmix.c
@@ -273,7 +273,7 @@ static void beg_activated(snd_state *ss)
     {
       mix_id = current_mix_id(ss);
       cp = mix_channel_from_id(mix_id);
-      set_mix_position_from_id(mix_id, (int)(string2Float(val) * SND_SRATE(cp->sound)));
+      set_mix_position_from_id(mix_id, (off_t)(string2Float(val) * SND_SRATE(cp->sound)));
       update_mix_panel(mix_id);
       XtFree(val);
     }
